refactor: setting 패키지 추가
