+++ src/main/java/Application.java
@@ -1,4 +1,3 @@
-import domain.game.JanggiGame;
 import infrastructure.BoardRepository;
 import infrastructure.DbConnection;
 import infrastructure.TurnRepository;
@@ -9,7 +8,7 @@ public class Application {
 
     public static void main(String[] args) {
         DbConnection dbConnection = new DbConnection();
-        
+
         JanggiGame janggiGame = new JanggiGame(
                 new InputView(), new OutputView(),
                 new BoardRepository(dbConnection), new TurnRepository(dbConnection)
