+++ src/main/java/domain/game/Turn.java
@@ -1,7 +1,6 @@
 package domain.game;
 
 import domain.piece.Country;
-import java.util.function.Consumer;
 
 public class Turn {
 
@@ -9,17 +8,6 @@ public class Turn {
 
     public Turn(Country current) {
         this.current = current;
-    }
-
-    public <T> void take(T value, Consumer<T> consumer) {
-        while (true) {
-            try {
-                consumer.accept(value);
-                return;
-            } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
-            }
-        }
     }
 
     public void next() {
