+++ src/test/java/domain/board/setting/strategy/MaSangMaSangTest.java
@@ -1,4 +1,4 @@
-package domain.board.strategy;
+package domain.board.setting.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,21 +8,21 @@ import domain.piece.PieceType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class MaSangSangMaTest {
+public class MaSangMaSangTest {
 
-    @DisplayName("장기판을 마상상마로 세팅한다.")
+    @DisplayName("장기판을 마상마상으로 세팅한다.")
     @Test
-    void maSangSangMaTest() {
-        Board board = new Board(new MaSangSangMa(), new MaSangSangMa());
+    void maSangMaSangTest() {
+        Board board = new Board(new MaSangMaSang(), new MaSangMaSang());
 
         assertThat(board.findPieceByCoordinate(new Coordinate(1, 8)).getType()).isEqualTo(PieceType.MA);
         assertThat(board.findPieceByCoordinate(new Coordinate(1, 7)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(1, 3)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(1, 2)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(1, 3)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(1, 2)).getType()).isEqualTo(PieceType.SANG);
 
         assertThat(board.findPieceByCoordinate(new Coordinate(10, 2)).getType()).isEqualTo(PieceType.MA);
         assertThat(board.findPieceByCoordinate(new Coordinate(10, 3)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(10, 7)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(10, 8)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(10, 7)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(10, 8)).getType()).isEqualTo(PieceType.SANG);
     }
 }
