+++ src/test/java/domain/board/setting/strategy/SangMaMaSangTest.java
@@ -1,4 +1,4 @@
-package domain.board.strategy;
+package domain.board.setting.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,21 +8,22 @@ import domain.piece.PieceType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class SangMaSangMaTest {
+public class SangMaMaSangTest {
 
-    @DisplayName("장기판을 상마상마로 세팅한다.")
+
+    @DisplayName("장기판을 상마마상로 세팅한다.")
     @Test
-    void sangMaSangMaTest() {
-        Board board = new Board(new SangMaSangMa(), new SangMaSangMa());
+    void sangMaMaSangTest() {
+        Board board = new Board(new SangMaMaSang(), new SangMaMaSang());
 
         assertThat(board.findPieceByCoordinate(new Coordinate(1, 8)).getType()).isEqualTo(PieceType.SANG);
         assertThat(board.findPieceByCoordinate(new Coordinate(1, 7)).getType()).isEqualTo(PieceType.MA);
-        assertThat(board.findPieceByCoordinate(new Coordinate(1, 3)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(1, 2)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(1, 3)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(1, 2)).getType()).isEqualTo(PieceType.SANG);
 
         assertThat(board.findPieceByCoordinate(new Coordinate(10, 2)).getType()).isEqualTo(PieceType.SANG);
         assertThat(board.findPieceByCoordinate(new Coordinate(10, 3)).getType()).isEqualTo(PieceType.MA);
-        assertThat(board.findPieceByCoordinate(new Coordinate(10, 7)).getType()).isEqualTo(PieceType.SANG);
-        assertThat(board.findPieceByCoordinate(new Coordinate(10, 8)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(10, 7)).getType()).isEqualTo(PieceType.MA);
+        assertThat(board.findPieceByCoordinate(new Coordinate(10, 8)).getType()).isEqualTo(PieceType.SANG);
     }
 }
